requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylint==2.3.1
pylint-plugin-utils==0.5
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
